+++ app/audio/analyzers/meter.py
@@ -48,4 +48,4 @@ class MeterAnalyzer(BaseAnalyzer):
             return {"meter": None}
 
     def _extract(self, ctx: AnalysisContext) -> dict[str, Any]:
-        return self.analyze(ctx.samples, ctx.sr)
+        return self.analyze(ctx.samples.astype(np.float32, copy=False), ctx.sr)
